requirements.txt
fastapi
uvicorn
jinja2
sentence-transformers
faiss-cpu
transformers
torch
wikipedia
python-dotenv
